+++ lib/radar.rb
@@ -32,6 +32,7 @@ module Radar
     autoload :BacktraceMatcher,    'radar/matchers/backtrace_matcher'
     autoload :ClassMatcher,        'radar/matchers/class_matcher'
     autoload :LocalRequestMatcher, 'radar/matchers/local_request_matcher'
+    autoload :MultiMatcher,        'radar/matchers/multi_matcher'
   end
 
   class Reporter