+++ lib/radar/matchers/backtrace_matcher.rb
@@ -3,9 +3,9 @@ module Radar
     # A matcher which matches exceptions which contain a certain
     # file in their backtrace.
     #
-    #     app.config.match :backtrace, "my_file"
-    #     app.config.match :backtrace, %r{lib/my_application}
-    #     app.config.match :backtrace, %r{lib/my_application}, :depth => 5
+    #     app.match :backtrace, "my_file"
+    #     app.match :backtrace, %r{lib/my_application}
+    #     app.match :backtrace, %r{lib/my_application}, :depth => 5
     #
     # By default this will search the entire backtrace, unless a depth
     # is specified.